+++ server/src/models/SixSidedDieClass.ts
